pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "openai>=1.61.1",
    "pillow>=11.1.0",
    "pyinstaller>=6.12.0",
    "sendgrid>=6.11.0",
]
